if else in delete in one go

+++ Jarvis/scripts/delete_in_one_go.sh
@@ -2,10 +2,17 @@
 
 result=$(find ~/.local/share/Trash/files/ -maxdepth 1 -iname "$1*")
 count=$(echo "$result" | wc -l)
-echo "$count result/s found to delete in one go"
+echo -n "Delete $count result/s in one go ? (yes/no): "
+read answer
 
-while read -r item; do
-    rm -rf "$item"
-done <<< "$result"
-
-echo "$count result/s deleted"
+if [ "$answer" == "yes" ]; then 
+    while read -r item; do
+        rm -rf "$item"
+    done <<< "$result"
+    echo "$count result/s deleted"
+elif [ "$answer" == "no" ]; then
+    echo "Deletion stopped"
+else
+    echo "Command Unclear. Deletion stopped"
+fi
+
